Add \rawx, \rawy, \rawz to let operation
There is an issue with 3D coordinates being projected onto the 2D canvas.
There is not way to reverse the projection because a system of two equations
(the project 2D coordinates) with three unknowns (the original 3D coordiantes)
is underdetermined and unsolvable.  Hence I introduce these new little helpers
which can be used to access the parsed original coordinates.  This works only
if you use literal coordinates in the let operation, i.e.

    \path let \p1=(1,2,3) in ...

will work as expected, but

    \coordinate (A) at (1,2,3);
    \path let \p1=(A) in ...

will not work, because the projection has already been performed when defining
the \coordinate.