Making the declared coordinate accessible
Often it is desirable to know what was the coordinate that a node (or other
named thing) was declared with.  This is now made globally available and
integrated into the let operation under the name \coord.  Keep in mind that
when calling \coord on something that was not a named thing, you'll just get
the declared coordinate of the last named thing or \relax if there was nothing
named before.  [[ Example:

    \path (0,0,{cos(45)}) coordinate (A);
    \path let \p1=(A), \p2=(2,2,2) in \pgfextra{\typeout{A=\coord1,B=\coord2}};

will print in the log

    A=(0,0,{cos(45)}),B=(0,0,{cos(45)})

because \p2 is not a named coordinate. ]]

Another issue might arise if the shape was not declare with a coordinate at
all, in which case the result is undefined.  You might end up with the name of
the thing itself or \relax. [[ Example:

    \coordinate (A);
    \path let \p1=(A) in \pgfextra{\typeout{A=\coord1}};

will print in the log

    A=(A)

because \p1 was not declare with a coordinate. ]]