More accurate \pgfpointnormalised #518 #519 Feature #96
This has been a long-standing problem, that \pgfpointnormalised was too
inaccurate.  This resulted in wrong intersection points when using the
projection syntax.  One proposal was to replace the definition of
\pgfpointnormalised by atan2 and cos and sin.  The prime argument against this
was the possible performance impact.  This is a reasonable assumption
considering the large definitions of atan2, cos, and sin.  However, in
typesetting the manual with the new and the old definition I could only observe
a 1% performance loss, which corresponds to 4s wallclock time on my machine.
Since the manual has an enormous amount of TikZ figures on over 1000 pages, I
think this performance loss can be considered negligible.