Stripping comments was too greedy
The old rule was

    \n%%[^\n]*\n

Which matched a newline followed by '%', followed by everything that is
not a newline, followed by another newline.  The trailing newline was
the culprit because it consumed the trailing newline, making a comment
on the next line not a match.  This

    % hello
    % world
    % foo
    % bar

transformed into this

    % world % bar

because of consuming the trailing newline.  Just leaving it there, i.e.

    \n%%[^\n]*

fixes the problem.