If prefixed name does not exist, look up global name #846
We had similar things already.  Usually, when a prefixed name does not
exist TikZ either bails or just does nothing.  Here we have the latter
kind.  If the prefixed node has no anchor, TikZ just does nothing and
leaves \tikz@shapeborder@name unset.  This has adverse effects when
drawing edges, because they store the node name and restore it later.
When the node name is restored, another name prefix is added to the
already prefixed node that was stored.  To remedy this, we now try the
unprefixed node, when the prefixed node does not exist.  This is
probably not 100% compatible but whoever was relying on this behavior
was on thin ice anyway.  The new behavior is more of the kind of “do
what I mean”.