New key `/pgf/fpu/install only' #861
This key will selectively install the specified functions from the fpu
library to be able to avoid `Dimension too large' errors in specific
cases.  Be aware that under certain circumstances the fpu comes with a
loss of precision due to the underlying fixed-point arithmetic of TeX.
These situations are rare but have be looked out for.

[[ Example: In general the fpu increases the range of permissible number
            substantially but sometimes has precision problems.

    \usetikzlibrary{fpu}
    \pgfkeys{/pgf/fpu/install only={divide}}
    \pgfmathparse{12.34/0.001234}\pgfmathresult % 10000.000000 (good)
    \pgfmathparse{12/4}\pgfmathresult % 2.9999000000000 (bad)

]]