transform shape clashes with label position #843
In the old code the conditional `\iftikz@label@simple' was set to true
when `transform shape' was in place. I do not understand the reason for
this at all. Honestly I also don't really understand large parts of the
rest of the code. It seems as if `\iftikz@label@simple' will only ever
be true when the label anchor coincides with the center anchor, which is
only ever true for nodes of type coordinate.