Correct documentation for .code 2 args second arg
The current documentation states that the key handler `.code 2 args`
treats the second argument as optional, but the implementation
differs: if the second argument isn't provided, then the first letter
of the first argument is treated as the first argument, and the
remainder of the first argument is treated as the second argument.
Thus, this change removes the incorrect statement from the
documentation.

Closes #991