Simplify \coordinate at parsing #785
This is another attempt at fixing #785 to make the \coordinate at
construction compatible with the (as of yet) internal \tikz@dcl@coord@
mechanism.

In some ways this is a quick and dirty fix because it doesn't remove
the flaw of the original implementation of expecting a literal ( at
the start of the coordinate. In principle the same as for
\tikz@scan@one@point should be done which expands tokens until it finds
a literal (.