Add reference for transform key
The description of the `text along path' decoration mentions using the
`transform` key to move text vertically (relative to the baseline).
Unfortunately, "transform" appears more than 1000 times in the
documentation, which makes finding the relevant documentation for that
key difficult, as evidenced personally and a question on TeX - LaTeX
Stack Exchange that addresses this issue.

This change adds a reference to the appropriate section in the
documentation for the `/pgf/decoration/transform` key so that future
users don't have this same difficulty. (It effectively follows the
recommendation in the answer to the TeX - LaTeX Stack Exchange
question, which has not been done in the intervening eight years.)

How can I move a "text along path" vertically? (TeX - LaTeX Stack
Exchange): https://tex.stackexchange.com/a/134589