fix!(pgfkeys): more careful treatment of \relax in pgfkeys
There are a lot of instances in pgfkeys where \csname is used, e.g. in
\pgfkeysgetvalueof. This will \let the mentioned csname become \relax so we have
to be prepared for this in across the board in pgfkeys. Indeed, this was the
original behavior before eTeX but it somehow got lost when transitioning to
\ifcsname. The behavior is broken with eTeX since 641a8832.

BREAKING CHANGE: \pgfkeysvalueof will expand to an alias for the primitve \relax
in case the key was undefined rather than to a \relaxed csname.

Co-authored-by: Jonathan Spratte <jspratte@yahoo.de>
Co-authored-by: muzimuzhi <muzimuzhi@gmail.com>
Signed-off-by: Henri Menke <henri@henrimenke.de>