Implement and document pgfkeys tracing.
The tracing code is based on the draft of package
`trace-pgfkeys.sty` (2012/02/06) by Ryan Reich <ryan.reich@gmail.com>.

`trace-pgfkeys` had a form of a runtime patch, and was thus difficult to
maintain.  The present implementation integrates the tracing code into
`pgfkeys.code.tex`.  The approach is DocStrip-like: every line of tracing code
occurs behind character `&`, which is set to either catcode 9 (ignore) or 14
(comment) to include or exclude the tracing code, respectively.

Consequently, the tracing code is not loaded unless specifically requested, and
therefore does not slow down normal key processing.

Signed-off-by: Sašo Živanović <saso.zivanovic@guest.arnes.si>