Update pgfmanual-en-base-shadings.tex
Line 258: missing `roll` operator (see line 301 where it is used). Line 288: probably *components* of colors, not coordinates of colors, for exemple components of the color yellow is 1 1 0 in rgb. Lines 271,272, not modified but `cvr` don't fix the broke in Apple Preview, see https://tex.stackexchange.com/q/739653/132405. Need to be updated if any solution is find.